requirements.txt
Django
gunicorn
whitenoise
psycopg2
websocket-client
python-binance
